requirements.txt
ddgs>=9.0.0
groq>=0.5.0
google-auth>=2.20.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
requests>=2.28.0